swap e 2opt feitos
ambas vizinhanças swap e 2opt estão feitas e estão funcionando de
maneira eficiente. agora, elas apenas simulam a operação realizada em
cada uma das iterações utilizando funções da classe solução que simulam
os movimentos da própria vizinhança.
aparentemente está funcionando tudinho

+++ include/instancia_problema.hpp
@@ -12,7 +12,7 @@ using namespace std;
 
 class instancia_problema {
 public:
-    int size;
+    ulong size;
     vector<suco_t> sucos;
     prepararLinha trocaSuco;
